[fix] Add AbortController for async upload operations (#12539)
## Describe your changes
This pull request improves the robustness of the `AudioInput` component
by adding support for aborting ongoing audio uploads. This ensures that
if a new upload is started or the component is unmounted, any previous
or ongoing upload is properly cancelled, preventing race conditions and
unnecessary state updates. (And the "tried to update state on an
unmounted component" error.)

<!-- If it's a visual change, please include a screenshot or video! -->

## GitHub Issue Link (if applicable)

No filed issue.

## Testing Plan

We could I guess try to repro this behavior in the tests, but since this
is a preventative measure, I don't think the ROI is worth adding a test
for this.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.