Ensure entry CSS file is index.css (#12780)
## Describe your changes

From 1.47 to 1.48, an index.css caused a conflict that made the entry
file not index.css. We enforce this in the vite config by detecting the
file and ensuring the name is index.css and the other css files contain
a hash.

## Testing Plan

- Automated Tests should pass.
- Verified use case of index.css will work
- Verified use case of index.css on local development

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.