[test] Update test paradigm for pdf with CCv2 (#13016)
## Describe your changes

Updated the PDF tests to assert against the CCv2 API instead of the CCv1
API. This is an update to
https://github.com/streamlit/streamlit/pull/13012. This change:

- Modified the `discover_and_register_components` method to accept a
`start_file_watching` parameter, allowing tests to disable file watching
- Updated the PDF tests to work with the CCv2 system
- Removed the version upper bound constraint for `streamlit-pdf` in
`test-requirements.txt`
- Enhanced the `DeltaGeneratorTestCase` to initialize and reuse a
`BidiComponentManager` for test. This matches the same lifecycle as
exists in the actual Runtime

## Testing Plan

- Unit Tests: The existing PDF tests have been updated to work with the
CCv2 framework

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.