Lower the asset size threshold (#13117)
## Describe your changes

With the recent improvements to asset and bundle size, we can lower the
loaded asset size threshold in our tests to 3.8MB. Currently, the actual
value is ~3.5 MB, but I think it's ok to have a bit of space to grow.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.