[feature] agent-knowledge processes (#12997)
## Describe your changes

Adds the `processes` directory with some documentation of things related
to creating PRs to support agents with this workflow.

For details of the purpose of this check out the README.

<!-- If it's a visual change, please include a screenshot or video! -->

## GitHub Issue Link (if applicable)

## Testing Plan

No tests necessary.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.