Activate pytest timeouts for e2e tests (#13201)
## Describe your changes

It seems like that sometimes individual tests seem to take a very long
time, causing the e2e workflow to hit its 30 min timeout. This is adding
a timeout of 300s that is applied to all of the individual e2e tests to
get better insights into which test cases are causing trouble.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.