Support for markdown in select_slider options (#12960)
## Describe your changes
I altered the slider_select to allow for markdown rendering in the
options field.

Now this,

`color = st.select_slider(
    "Select a color of the rainbow",
    options=[
        "*red*",
        "*orange*",
        "yellow",
        "green",
        "blue",
        "indigo",
        "**violet**",
    ],
)`

kind of code renders correctly like in the picture.
Not just as `*red*` but actually as *red*.

## Screenshot or video (only for visual changes)
<img width="778" height="177" alt="Untitled"
src="https://github.com/user-attachments/assets/2a146fa4-a132-477d-8ea2-b7e9bf0c4e16"
/>

## GitHub Issue Link (if applicable)
https://github.com/streamlit/streamlit/issues/11774

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
- E2E Tests
- Any manual testing needed?

Further testing would be useful but I am not sure if the issue is still
being pursued, as it is older. If yes I can do that and also update the
docs.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: Bob Nisco <bob.nisco@snowflake.com>