Automate ruff pre-commit version update (#13395)
## Describe your changes

Whenever ruff gets updated (by dependabot), we also need to manually
update to the same version in the pre-commit config. This is a small but
annoying task. Adding an automation script for this to precommit, which
will be auto-fixed by our autofix workflow whenever dependabot creates a
ruff update and ensures the versions are always in sync.




---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.