Allow dynamic changes to selectbox options when `key` is provided (#13383)
## Describe your changes

Allow dynamically changing the options for `st.selectbox` without
triggering an identity change / state reset. If the current selected
options isn't in the list of available option, it will be reset to the
default value.

## GitHub Issue Link (if applicable)

- Closes https://github.com/streamlit/streamlit/issues/6352
- Closes https://github.com/streamlit/streamlit/issues/8496
- Closes https://github.com/streamlit/streamlit/issues/4854
- Related: https://github.com/streamlit/streamlit/issues/11277

## Testing Plan

- Added unit and e2e tests.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.