Expose OIDC tokens (#12044)
## Describe your changes

Store the OIDC auth response in a cookie and expose them to users for
API calls

Currently, from user perspective, `st.user` gets an extra field called
"tokens" that is a dict of all the tokens that were made available -
most often the triplet of acces_token, id_token, refresh_token

This is useful for multiple use cases (see the issues below).

To achieve (a) reliably, the tokens are stored in a separate cookie from
user info. This is because cookies are limited to be <4096 bytes and I
ran into that limit in my own testing already if I combined the two
dictionaries. Additionally, both cookies can now be chunked so as a
by-product, this now supports login systems with larger userdata.

I have tested it to work in my own use case and it seems to be
performing as intended. But it is just one use case and I may well be
missing some others. so any proposals to improve this are very welcome.

Originally, also contained st.user_refresh command, but that got it's
own PR: https://github.com/streamlit/streamlit/pull/12696

## GitHub Issue Link (if applicable)

Closes #10378 

Exposing access_token:
https://github.com/streamlit/streamlit/issues/10378
Too large usedata for current login system:
https://github.com/streamlit/streamlit/issues/12518

Need to provide id token hint at logout:
https://github.com/streamlit/streamlit/issues/12144#issuecomment-3172828767
-> PR https://github.com/streamlit/streamlit/pull/12693
Need for user refresh:
https://github.com/streamlit/streamlit/issues/12043 -> PR
https://github.com/streamlit/streamlit/pull/12696

## Testing Plan

- Unit tests present
- Manually tested to work for my own use case

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: Ken McGrady <ken.mcgrady@gmail.com>